+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Don't count dormant minion attack (issue #1234)
 - Fixed an issue where the battlegrounds overlay would not work with hero skins in some cases
 - Fixed an issue where questlines were not considered quests
+- Fixed an issue where HST did not count games played against an opponent with Maestra as games against a Rogue
 
 # 1.9.11
 ## New